+++ src/Controller/ArticleReferenceAdminController.php
@@ -101,6 +101,36 @@ class ArticleReferenceAdminController extends BaseController
 }
 
     /**
+     * Now we reordering all of refereneces after drag element in order list
+     * @Route("/admin/article/{id}/references/reorder", methods="POST", name="admin_article_reorder_references")
+     * @IsGranted("MANAGE", subject="article")
+     */
+    public function reorderArticleReferences(Article $article, Request $request, EntityManagerInterface $entityManager)
+    {
+        $orderedIds = json_decode($request->getContent(), true);
+        if ($orderedIds === null) {
+            return $this->json(['detail' => 'Invalid body'], 400);
+        }
+
+        // from (position)=>(id) to (id)=>(position)
+        $orderedIds = array_flip($orderedIds);
+
+        foreach ($article->getArticleReferences() as $reference) {
+            $reference->setPosition($orderedIds[$reference->getId()]);
+        }
+        $entityManager->flush();
+
+        return $this->json(
+            $article->getArticleReferences(),
+            200,
+            [],
+            [
+                'groups' => ['main']
+            ]
+        );
+    }
+
+    /**
      * API endpoint to get current list of references
      * @Route("/admin/article/{id}/references", methods="GET", name="admin_article_list_references")
      * @IsGranted("MANAGE", subject="article")